+++ ai/common/game_state_helpers.py
