+++ dummy_player.py
@@ -1,0 +1,1 @@
+from ai.agents.dummy import Player
