+++ minimax_player.py
@@ -1,0 +1,1 @@
+from ai.agents.minimax import Player
